cleanup markdown
