--- lua/plugins/configs/markdown/bullets.lua
@@ -1,8 +1,0 @@
-local M = {}
-
-M.setup = function()
-	vim.cmd([[
-  let g:bullets_enabled_file_types = ['markdown', 'text', 'gitcommit', 'scratch']
-  ]])
-end
-return M
