--- lua/plugins/configs/markdown/md_snippets.lua
@@ -1,26 +1,0 @@
-vim.cmd([[
-    autocmd Filetype markdown inoremap <buffer> ,f <++>
-    autocmd Filetype markdown inoremap <buffer> ,n ---<Enter><Enter>
-    autocmd Filetype markdown inoremap <buffer> ,b **** <++><Esc>F*hi
-    autocmd Filetype markdown inoremap <buffer> ,sb <sub></sub> <++><Esc>Fs2hi
-    autocmd Filetype markdown inoremap <buffer> ,sp <sup></sup> <++><Esc>Fs2hi
-    autocmd Filetype markdown inoremap <buffer> ,i ** <++><Esc>F*i
-    autocmd Filetype markdown inoremap <buffer> ,d `` <++><Esc>F`i
-    autocmd Filetype markdown inoremap <buffer> ,c ```<Enter><++><Enter>```<Enter><Enter><++><Esc>4kA
-    autocmd Filetype markdown inoremap <buffer> ,m - [ ] 
-    autocmd Filetype markdown inoremap <buffer> ,p ![](<++>) <++><Esc>F[a
-    autocmd Filetype markdown inoremap <buffer> ,a [](<++>) <++><Esc>F[a
-    autocmd Filetype markdown inoremap <buffer> ,1 #<Space><Enter><++><Esc>kA
-    autocmd Filetype markdown inoremap <buffer> ,2 ##<Space><Enter><++><Esc>kA
-    autocmd Filetype markdown inoremap <buffer> ,3 ###<Space><Enter><++><Esc>kA
-    autocmd Filetype markdown inoremap <buffer> ,4 ####<Space><Enter><++><Esc>kA
-    autocmd Filetype markdown inoremap <buffer> ,l --------<Enter>
-    autocmd Filetype markdown inoremap <buffer> ,l --------<Enter>
-    autocmd Filetype markdown inoremap <buffer> ,w <details><Enter><summary></summary><Enter><Enter><++><Enter></details><Esc>3ki
-
-  ]])
-
--- autocmd Filetype markdown map <leader>w yiWi[<esc>Ea](<esc>pa)
--- autocmd Filetype markdown inoremap <buffer> ,f <Esc>/<++><CR>:nohlsearch<CR>"_c4l
--- autocmd Filetype markdown inoremap <buffer> ,w <Esc>/ <++><CR>:nohlsearch<CR>"_c5l<CR>
--- autocmd Filetype markdown inoremap <buffer> ,s ~~~~ <++><Esc>F~hi
